requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
gspread
google-auth
